fixed string separator unit tests
