+++ System/tests/Extensions/Strings/StringSelectorTests.cs
@@ -1,6 +1,5 @@
 ﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
 
-#nullable enable
 
 using Wangkanai.Exceptions;
 
@@ -11,13 +10,13 @@ namespace Wangkanai.Extensions.Strings;
 public class StringSelectorTests
 {
 	string? _null  = null;
-	string? _empty = string.Empty;
-	string? _space = " ";
+	string _empty = string.Empty;
+	string _space = " ";
 
 	[Fact]
 	public void Null()
 	{
-		Assert.Throws<ArgumentNullException>(() => _null.EscapeSelector());
+		Assert.Throws<ArgumentNullException>(() => _null!.EscapeSelector());
 	}
 
 	[Fact]
